refactor: split token ring into srp modules
