+++ crates/bitnet-common/src/token_ring/display.rs
@@ -1,0 +1,7 @@
+use super::TokenRing;
+
+impl std::fmt::Display for TokenRing {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "TokenRing({}/{})", self.len, self.capacity)
+    }
+}
