+++ crates/bitnet-tool-use-core/src/parsing.rs
@@ -54,12 +54,15 @@ mod tests {
     use super::*;
 
     #[test]
-    fn parse_tool_call_valid() {
+    fn parse_tool_call_valid() -> Result<(), Box<dyn std::error::Error>> {
         let text =
             r#"<tool_call>{"name":"get_weather","arguments":{"location":"London"}}</tool_call>"#;
-        let call = parse_tool_call(text, &ToolUseFormat::ChatMLTools).expect("valid tool call");
+        let Some(call) = parse_tool_call(text, &ToolUseFormat::ChatMLTools) else {
+            return Err("expected valid ChatML tool call".into());
+        };
         assert_eq!(call.name, "get_weather");
         assert!(call.arguments.contains("London"));
+        Ok(())
     }
 
     #[test]
